extract separate utils/ modules
