+++ src/utils/tuple.tsx
@@ -1,0 +1,3 @@
+const tuple = <T extends any[]>(...args: T): T => args
+
+export default tuple
